Conversion complete; New layout; Fix #24; Fix #32;
