+++ src/chooser.html
@@ -1,0 +1,22 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<title>CrossOver</title>
+		<link rel="stylesheet" href="index.css">
+	</head>
+	<body>
+		<div id="chooser-container">
+
+			<div id="chooser-wrapper">
+
+				<div id="crosshair-chooser"></div>
+
+			</div>
+
+		</div>
+
+		<script src="./chooser.js"></script>
+
+	</body>
+</html>
