+++ src/index.scss
@@ -29,6 +29,13 @@ body {
 	}
 }
 
+button {
+	cursor: pointer;
+	border: none;
+	outline: none;
+	background-color: transparent;
+}
+
 .help-text {
 	text-align: center;
 	letter-spacing: 1.1px;
@@ -56,6 +63,23 @@ body {
 	color: inherit;
 }
 
+// Blue "Hover" border
+.container,
+#settings-container,
+#chooser-container {
+	&.dropping:before {
+		content: '';
+		position: absolute;
+		top: 5px;
+		right: 5px;
+		bottom: 5px;
+		left: 5px;
+		border: 4px dashed $border_color_focus;
+		pointer-events: none;
+		z-index: 9;
+	}
+}
+
 /* Icon Styles */
 .feather {
 	// stroke:whitesmoke;
@@ -65,7 +89,6 @@ body {
 
 /* Application Style */
 .container {
-	padding: 15px 4px 4px 4px;
 	overflow: hidden;
 	position: relative;
 	background: none;
@@ -74,12 +97,6 @@ body {
 	height: 100%;
 	-webkit-user-select: none;
 	cursor: none;
-
-	&.dropping {
-		// box-shadow: inset 0 0 10px $border_color_focus;
-		padding: 11px 0 0 0;
-		border: 4px dashed $border_color_focus;
-	}
 }
 
 .background {
@@ -93,28 +110,34 @@ body {
 	z-index: -1;
 }
 
-.center-me {
+.button {
 	display: none;
 	position: absolute;
-	top: 1px;
-	left: 0;
 	-webkit-user-select: none;
 	z-index: 2;
-	filter: drop-shadow(0px 0px 3px rgba(0,0,0,0.5));
 	.feather {
-		stroke:rgba(255,255,255,0.6);
+		stroke:rgba(0,0,0,0.5);
 	}
 }
 
-.drag-me {
-	display: none;
-	position: absolute;
-	top: 1px;
-	right: 0;
-	-webkit-user-select: none;
-	z-index: 2;
+.settings-button {
+	bottom: 1px;
+	right: 1px;
+}
+
+.center-button {
+	bottom: 1px;
+	left: 1px;
+	// filter: drop-shadow(0px 0px 3px rgba(0,0,0,0.5));
 	.feather {
-		// stroke:rgba(255,255,255,0.8);
+		// stroke:rgba(255,255,255,0.6);
+	}
+}
+
+.drag-button {
+	top: 2px;
+	right: 2px;
+	.feather {
 		width: 12px;
 		height: 12px;
 	}
@@ -129,10 +152,12 @@ body {
 }
 
 .crosshair-wrapper {
+	// Top is draggable
+	padding-top: 25px;
+	margin-bottom: 25px;
 	position: relative;
 	height: 100px;
 	width: 100%;
-	padding-bottom: 15px;
 	display: flex;
 	align-items: center;
 	justify-content: center;
@@ -163,38 +188,46 @@ body {
 }
 
 .draggable {
+	.crosshair-wrapper,
+	.crosshair-wrapper * {
+		// This is the magic to make things draggable
+		-webkit-app-region: drag;
+	}
+
 	.container {
-		cursor: inherit;
+		cursor: crosshair;
+		border-bottom-right-radius: 0;
+		border-bottom-left-radius: 0;
 	}
-	&:not(.pickr-open) {
-		.crosshair-wrapper,
-		.crosshair-wrapper * {
-			// This is the magic to make things draggable
-			-webkit-app-region: drag;
-		}
-	}
+
 	.background {
 		background: rgba(255, 255, 255, 0.5);
 	}
-	.drag-me {
+
+	.button {
 		display: block;
 	}
-	.center-me {
-		// Allow center on double-click (Windows; https://github.com/lacymorrow/crossover/issues/32)
-		-webkit-app-region: no-drag;
-		display: block;
-		cursor: crosshair;
 
-	}
-	.settings {
-		display: flex;
-	}
-
-	&.mac:not(.pickr-open) .crosshair-wrapper * {
+	.drag-button {
 		cursor: grab;
 		&:active {
 			cursor: grabbing;
 		}
+	}
+
+	.center-button {
+		// Allow center on double-click (Windows; https://github.com/lacymorrow/crossover/issues/32)
+		-webkit-app-region: no-drag;
+		display: block;
+
+	}
+
+	.settings {
+		display: flex;
+	}
+
+	&.mac .crosshair-wrapper {
+		cursor: crosshair;
 	}
 }
 
@@ -394,33 +427,34 @@ input[type='range'] {
 }
 
 #select-crosshair-button {
-	cursor: pointer;
-	border: none;
-	outline: none;
+	padding-top: 5px;
+	padding-left: 20px;
 	font-size: 14px;
 	font-weight: 700;
 	text-decoration: underline;
 	color: #2466d8;
-	background-color: transparent;
 }
 
 /* Crosshair chooser */
 
+
+#chooser-container,
 #settings-container {
 	height: 100vh;
 	width: 100vw;
-	background: transparent;
-	padding: 4px 4px 4px 4px;
-
-	&.dropping {
-		// box-shadow: inset 0 0 10px $border_color_focus;
-		padding: 0 0 0 0;
-		border: 4px dashed $border_color_focus;
-	}
+	padding: 4px;
+	background: white;
 }
 
+#settings-container {
+}
+
+#chooser-container {
+	border-radius: 5px;
+}
+
+#chooser-wrapper,
 #settings-wrapper {
-	background: white;
 	height: 100%;
 }
 
@@ -464,10 +498,11 @@ input[type='range'] {
 		display: block;
 		padding: 0 0.5em;
 		font-size: 20px;
-		color: #2466d8;
+		// color: #2466d8; // Blue
 		margin-bottom: 0;
 	}
 
+	// Collection wrapper
 	&>div {
 		width: 100%;
 		padding: 0 0.5em 0;
@@ -476,7 +511,11 @@ input[type='range'] {
 		display: flex;
 		flex-direction: row;
 		flex-wrap: wrap;
-		align-items: center;
+		align-items: start;
+
+		p {
+			max-width: 88px;
+		}
 	}
 
 	img {
